requirements.txt
google-cloud-storage
gcsfs
waymo-open-dataset-tf-2-6-0
